Rename Module: cdsw to cdp
