+++ rdsa_utils/cdp/helpers/impala.py
